Reorganize the old command line code to get ready for the new CLI.
The only functionality change is that "--verbose" is no longer a
supported option. It was only used for help, and this just always shows
all the help now.

The other changes mostly move code out of bin/format.dart so that
relevant pieces can be shared with the new CLI.
